requirements.txt
accelerate==0.27.2
adaseq==0.6.6
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
av==11.0.0
Babel==2.13.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.2
bitsandbytes==0.37.0
bleach==6.1.0
blis==0.7.11
boto3==1.34.56
botocore==1.34.56
bpemb==0.3.4
braceexpand==0.1.7
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cached-path==1.1.6
cachetools==5.3.3
catalogue==2.0.10
cchardet==2.1.7
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.1.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.0
commonmark==0.9.1
confection==0.1.4
conllu==4.5.3
contourpy==1.0.7
crcmod==1.7
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.11.0
cymem==2.0.8
Cython==3.0.8
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.6
fastapi==0.104.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.7.1
flair==0.13.1
flash-attn==1.0.7
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fschat==0.2.36
fsspec==2023.12.1
ftfy==6.1.3
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==5.1.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gradio==3.24.1
gradio_client==0.0.8
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.21.4
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
importlib-metadata==7.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.27.1
ipython==8.12.3
isoduration==20.11.0
Janome==0.5.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonnet==0.20.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
linkify-it-py==2.0.2
llvmlite==0.41.1
lmdb==1.4.1
lxml==5.1.0
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
modelscope==1.12.0
more-itertools==10.2.0
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy @ file:///croot/numpy_and_numpy_base_1701295038894/work/dist/numpy-1.26.2-cp39-cp39-linux_x86_64.whl#sha256=ab2439928d6a64e481fc12c271cde207be02e933033d4fecde57d3e534b5025c
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.0
opencv-python==4.7.0.72
orjson==3.9.10
oss2==2.18.4
overrides==7.4.0
packaging==23.0
pandas==2.1.4
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.3
peft
pexpect==4.9.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1696580024257/work
pipreqs==0.5.0
platformdirs==4.1.0
pluggy==1.4.0
portalocker==2.8.2
pptree==3.1
preshed==3.0.9
prettytable==3.10.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydantic==1.8.2
pydub==0.25.1
Pygments==2.17.2
pylocker==3.1.0
pynndescent==0.5.11
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
pyrep==3.2.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytest==8.0.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-crf==0.7.2
pytorch_revgrad==0.2.0
pytorchvideo==0.1.5
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.2
referencing==0.32.0
regex==2022.10.31
replicate==0.24.0
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.13.2
rsa==4.9
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
scripts==3.0
segtok==1.5.11
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.38.0
seqeval==1.2.2
setproctitle==1.3.3
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.3.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
stua==0.3
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboardX==2.6.2.2
termcolor==1.1.0
terminado==0.18.0
thinc==8.0.17
threadpoolctl==3.2.0
tifffile==2024.2.12
tiktoken==0.6.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.64.1
traitlets==5.14.0
transformer-smaller-training-vocab==0.3.3
transformers==4.21.0
triton==2.1.0
tweepy==4.14.0
typer==0.4.2
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.5
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.24.0.post1
wandb==0.12.21
wasabi==0.10.1
wavedrom==2.0.3.post3
wcwidth==0.2.12
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Wikipedia-API==0.6.0
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
yarl==1.8.2
zipp==3.14.0
